2021 Day 15 Remove redundant visit check

+++ 2021/15/main_test.go
@@ -16,8 +16,8 @@ var safeinput []byte
 type TestCase struct {
 	file string
 	data []byte
-	dim int
-	ans  int
+	dim uint16
+	ans  uint16
 }
 
 func TestSolve(t *testing.T) {
